backend/requirements.txt
Django==5.0.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
mysqlclient==2.2.0
Pillow==10.0.0
python-dotenv==1.0.0
django-filter==23.2
